requirements.txt
fastapi
uvicorn
requests
pdf2image
google-generativeai
pydantic
python-dotenv
pillow